Initial startup
spent time to understand the data and what to stratify. Stratified "step" becuase there are alot less entries in steps 1-7 then there are in 7-13. So sampling 1-7 more than 7-13 would be very biased and innacurate. Playing around with plots and following the example to see what the code does and how to use it.